sac-her trainer merged to single-step-trainer-update
