+++ requirements.txt
@@ -1,3 +1,8 @@
+# for stable_baselines3
 stable-baselines3[extra,tests,docs]>=1.1.0a11
 pybullet
 sb3-contrib>=1.1.0a11
+
+# formatter
+black
+isort
